pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "zombie-hunter"
version = "0.1.0"
description = "A FinOps tool that hunts zombie cloud resources and enables cleanup via Slack"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "Zombie Hunter Team"}
]
keywords = ["finops", "cloud", "aws", "gcp", "azure", "kubernetes", "cost-optimization"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Systems Administration",
]

dependencies = [
    "boto3>=1.34.0",
    "google-cloud-compute>=1.14.0",
    "google-cloud-monitoring>=2.18.0",
    "azure-identity>=1.15.0",
    "azure-mgmt-compute>=30.0.0",
    "azure-mgmt-network>=25.0.0",
    "azure-mgmt-monitor>=6.0.0",
    "slack-bolt>=1.18.0",
    "slack-sdk>=3.26.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "pyyaml>=6.0.1",
    "structlog>=24.1.0",
    "rich>=13.7.0",
    "click>=8.1.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "moto>=5.0.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
    "types-pyyaml>=6.0.0",
    "boto3-stubs[ec2,rds,elbv2,cloudwatch]>=1.34.0",
]

[project.scripts]
zombie-hunter = "zombie_hunter.main:main"

[project.urls]
Homepage = "https://github.com/yourusername/zombie-hunter"
Repository = "https://github.com/yourusername/zombie-hunter"
Issues = "https://github.com/yourusername/zombie-hunter/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["zombie_hunter*"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
